Add files via upload
In this project, we started transforming the video images into grayscale images then we reduced the frame rate. After that, we created a video from a set of 2D images of pieton and tried to track only one pieton from the video. Finally, we created synthetic images with moving objects.